preinstall: only check for CPU debugging on Intel
The check for CPU silicon-level debugging is testing the contents of a
MSR that is Intel specific. When we have checks for AMD CPUs, this test
won't apply to that (or, we may need an AMD-specific test).
