+++ efi/preinstall/check_fw_protections_intel_test.go
@@ -27,6 +27,7 @@ import (
 
 	. "gopkg.in/check.v1"
 
+	"github.com/canonical/cpuid"
 	. "github.com/snapcore/secboot/efi/preinstall"
 	internal_efi "github.com/snapcore/secboot/internal/efi"
 	"github.com/snapcore/secboot/internal/efitest"
@@ -612,4 +613,52 @@ func (s *fwProtectionsIntelSuite) TestCheckPlatformFirmwareProtectionsMEIErrNoFP
 	err := CheckPlatformFirmwareProtectionsIntelMEI(env)
 	c.Check(err, ErrorMatches, `no hardware root-of-trust properly configured: BootGuard OTP fuses are not locked`)
 	c.Check(err, FitsTypeOf, &NoHardwareRootOfTrustError{})
+}
+
+func (s *fwProtectionsIntelSuite) TestCheckCPUDebuggingLockedMSRDisabledCPUID(c *C) {
+	env := efitest.NewMockHostEnvironmentWithOpts(efitest.WithAMD64Environment("GenuineIntel", nil, 1, map[uint32]uint64{0xc80: 0}))
+	amd64Env, err := env.AMD64()
+	c.Assert(err, IsNil)
+
+	c.Check(CheckCPUDebuggingLockedMSR(amd64Env), IsNil)
+}
+
+func (s *fwProtectionsIntelSuite) TestCheckCPUDebuggingLockedMSRDisabledMSR(c *C) {
+	env := efitest.NewMockHostEnvironmentWithOpts(efitest.WithAMD64Environment("GenuineIntel", []uint64{cpuid.SDBG}, 4, map[uint32]uint64{0xc80: 0x40000000}))
+	amd64Env, err := env.AMD64()
+	c.Assert(err, IsNil)
+
+	c.Check(CheckCPUDebuggingLockedMSR(amd64Env), IsNil)
+}
+
+func (s *fwProtectionsIntelSuite) TestCheckCPUDebuggingLockedMSRDisabledAvailable(c *C) {
+	env := efitest.NewMockHostEnvironmentWithOpts(efitest.WithAMD64Environment("GenuineIntel", []uint64{cpuid.SDBG}, 4, map[uint32]uint64{0xc80: 0}))
+	amd64Env, err := env.AMD64()
+	c.Assert(err, IsNil)
+
+	c.Check(CheckCPUDebuggingLockedMSR(amd64Env), Equals, ErrCPUDebuggingNotLocked)
+}
+
+func (s *fwProtectionsIntelSuite) TestCheckCPUDebuggingLockedMSREnabled(c *C) {
+	env := efitest.NewMockHostEnvironmentWithOpts(efitest.WithAMD64Environment("GenuineIntel", []uint64{cpuid.SDBG}, 4, map[uint32]uint64{0xc80: 1}))
+	amd64Env, err := env.AMD64()
+	c.Assert(err, IsNil)
+
+	c.Check(CheckCPUDebuggingLockedMSR(amd64Env), Equals, ErrCPUDebuggingNotLocked)
+}
+
+func (s *fwProtectionsIntelSuite) TestCheckCPUDebuggingLockedMSRErrMissingMSR(c *C) {
+	env := efitest.NewMockHostEnvironmentWithOpts(efitest.WithAMD64Environment("GenuineIntel", []uint64{cpuid.SDBG}, 4, map[uint32]uint64{}))
+	amd64Env, err := env.AMD64()
+	c.Assert(err, IsNil)
+
+	c.Check(CheckCPUDebuggingLockedMSR(amd64Env), ErrorMatches, `MSR does not exist`)
+}
+
+func (s *fwProtectionsIntelSuite) TestCheckCPUDebuggingLockedMSRErrNoMSRValues(c *C) {
+	env := efitest.NewMockHostEnvironmentWithOpts(efitest.WithAMD64Environment("GenuineIntel", []uint64{cpuid.SDBG}, 0, map[uint32]uint64{0xc80: 0x40000000}))
+	amd64Env, err := env.AMD64()
+	c.Assert(err, IsNil)
+
+	c.Check(CheckCPUDebuggingLockedMSR(amd64Env), ErrorMatches, `no MSR values returned`)
 }
